Add tabs to switch between leaderboard and tee times views
- Replace "Switch to" links with tab UI on both competition pages
- Extract CutInfo into its own component so it can be reused
- Show CutInfo on the tee times page by fetching leaderboard data
