+++ styles.css
@@ -130,8 +130,29 @@ button {
   opacity: 0.7;
 }
 
+.page-tabs {
+  display: flex;
+  gap: 4px;
+  padding: 8px 10px 10px;
+}
+
+.page-tab {
+  padding: 4px 12px;
+  border-radius: 4px;
+  font-size: 14px;
+  text-decoration: none;
+  color: inherit;
+  opacity: 0.6;
+}
+
+.page-tab--active {
+  background: var(--primary);
+  color: #fff;
+  opacity: 1;
+}
+
 .cut-info {
-  margin: 10px 10px 20px;
+  margin: 10px 10px 12px;
   border-left: 5px solid var(--primary);
   padding-left: 10px;
 }
